Feat(Auth): View에서 웹 이동 시 fetchWithRefresh 사용하도록 변경. 비활성 유저 로그인 불가하도록 기능 추가. 관광지, 숙소 버튼 디자인 etc..

+++ src/main/java/com/multi/travel/admin/service/AdminService.java
@@ -153,7 +153,7 @@ public class AdminService {
                 savedFileName = FileUploadUtils.saveFile(IMAGE_DIR, uniqueFileName, file);
 
                 // ✅ 기존 이미지 삭제 (URL → 파일명 변환)
-                if (oldImage != null && !oldImage.isEmpty()) {
+                if (!oldImage.startsWith("http://tong") && oldImage != null && !oldImage.isEmpty()) {
                     String oldFileName = oldImage.replace(IMAGE_URL, "");
                     FileUploadUtils.deleteFile(IMAGE_DIR, oldFileName);
                     log.info("[updateSpot] 기존 이미지 삭제: {}", oldFileName);
